new features added, few code optimizations
