+++ .env
@@ -3,3 +3,5 @@ DB_USER=postgres
 DB_PASSWORD=.GzPAj8-j;6R#FM0
 DB_NAME=postgres
 DB_PORT=5432
+API_SECRET=dima123
+TOKEN_HOUR_LIFESPAN=1
