+++ cmd/auth/auth.go
@@ -1,7 +1,3 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
-*/
 package auth
 
 import (
@@ -15,29 +11,15 @@ var (
 	userID uint
 )
   
-// authCmd represents the auth command
 var AuthCmd = &cobra.Command{
 	Use:   "auth",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "Authorization commands are set here",
+	Long: `Authorization commands are set here (login and signup)`,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("Call one of following commands: `signup` or `login`")
 	},
 }
 
 func init() {
-	// Here you will define your flags and configuration settings.
 
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// authCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// authCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
