+++ cmd/auth/login.go
@@ -1,10 +1,7 @@
-/*
-Copyright © 2023 NAME HERE <EMAIL ADDRESS>
-
-*/
 package auth
 
 import (
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"os"
@@ -13,7 +10,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// loginCmd represents the login command
 var loginCmd = &cobra.Command{
 	Use:   "login",
 	Short: "Login command to be able to interact with another commands.",
@@ -27,15 +23,8 @@ func init() {
 	AuthCmd.AddCommand(loginCmd)
 	loginCmd.Flags().StringVarP(&userName, "name", "n", "", "name of your user")
 	loginCmd.Flags().StringVarP(&userPassword, "password", "p", "", "your password")
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// loginCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// loginCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	loginCmd.MarkFlagRequired("name")
+	loginCmd.MarkFlagRequired("password")
 }
 
 func Login() {
@@ -52,25 +41,44 @@ func Login() {
 	  os.Exit(1)
 	}
 	db.Model(&utils.User{}).Where("name =? and password =?", userName, userPassword).Pluck("id", &userID)
-	UserRefer, err1 := json.MarshalIndent(userID, "", "/t")
-	if err1 != nil {
-		panic(err1)
+	UserRefer, err := json.MarshalIndent(userID, "", "/t")
+	if err != nil {
+		fmt.Println("Impossible to extract user ID")
 	}
-	os.WriteFile("UserID", UserRefer, 0644)
+	encodedStr := base64.StdEncoding.EncodeToString(UserRefer)
+
+	err1 := os.WriteFile("UserID", []byte(encodedStr), 0644)
+	if err1 != nil{
+		fmt.Println("Impossible to extract user ID")
+	}
 	fmt.Println("You are sucessfully logged in")
 }
 
 func CheckLogin() uint{
 
-	uid, err2 := os.ReadFile("UserID") 
-	if err2 != nil {
+	uid, err := os.ReadFile("UserID") 
+	if err != nil {
 		fmt.Println("You are not logged in. Please log in with `login` command")
 		os.Exit(1)
 	}else {
-		err3 := json.Unmarshal(uid, &userID)
-		if err3 != nil {
-			panic("Impossible to extract UserID from file")
+		err := json.Unmarshal(uid, &userID)
+		if err != nil {
+			fmt.Println("Impossible to extract UserID from file")
+			os.Exit(1)
+		}
+		idStr := string(userID)
+		decodedID, err := base64.StdEncoding.DecodeString(idStr)
+		if err != nil {
+			fmt.Println("Impossible to extract UserID from file")
+			os.Exit(1)
+		}
+		
+		err1 := json.Unmarshal(decodedID, &userID)
+		if err1 != nil {
+			fmt.Println("Impossible to extract UserID from file")
+			os.Exit(1)
 		}
 	}
+
 	return userID
 }
